+++ pkgs/ffigen/lib/src/header_parser/sub_parsers/typedefdecl_parser.dart
@@ -36,9 +36,13 @@ Typealias? parseTypedefDeclaration(
   final logger = context.logger;
   final config = context.config;
   final bindingsIndex = context.bindingsIndex;
-  final typedefName = cursor.spelling();
-  final typedefUsr = cursor.usr();
-  final decl = Declaration(usr: typedefUsr, originalName: typedefName);
+  final name = cursor.spelling();
+  final usr = cursor.usr();
+
+  final cachedType = bindingsIndex.getSeenTypealias(usr);
+  if (cachedType != null) return cachedType;
+
+  final decl = Declaration(usr: usr, originalName: name);
   final ct = clang.clang_getTypedefDeclUnderlyingType(cursor);
   final s = getCodeGenType(
     context,
@@ -47,46 +51,48 @@ Typealias? parseTypedefDeclaration(
     originalCursor: cursor,
   );
 
-  if (bindingsIndex.isSeenUnsupportedTypealias(typedefUsr)) {
+  if (bindingsIndex.isSeenUnsupportedTypealias(usr)) {
     // Do not process unsupported typealiases again.
   } else if (s is UnimplementedType) {
     logger.fine(
-      "Skipped Typedef '$typedefName': "
+      "Skipped Typedef '$name': "
       'Unimplemented type referred.',
     );
-    bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
-  } else if (s is Compound && s.originalName == typedefName) {
+    bindingsIndex.addUnsupportedTypealiasToSeen(usr);
+  } else if (s is Compound && s.originalName == name) {
     // Ignore typedef if it refers to a compound with the same original name.
-    bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    bindingsIndex.addUnsupportedTypealiasToSeen(usr);
     logger.fine(
-      "Skipped Typedef '$typedefName': "
+      "Skipped Typedef '$name': "
       'Name matches with referred struct/union.',
     );
   } else if (s is EnumClass) {
     // Ignore typedefs to Enum.
-    bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
-    logger.fine("Skipped Typedef '$typedefName': typedef to enum.");
+    bindingsIndex.addUnsupportedTypealiasToSeen(usr);
+    logger.fine("Skipped Typedef '$name': typedef to enum.");
   } else if (s is HandleType) {
     // Ignore typedefs to Handle.
-    logger.fine("Skipped Typedef '$typedefName': typedef to Dart Handle.");
-    bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    logger.fine("Skipped Typedef '$name': typedef to Dart Handle.");
+    bindingsIndex.addUnsupportedTypealiasToSeen(usr);
   } else if (s is ConstantArray || s is IncompleteArray) {
     // Ignore typedefs to Constant Array.
-    logger.fine("Skipped Typedef '$typedefName': typedef to array.");
-    bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    logger.fine("Skipped Typedef '$name': typedef to array.");
+    bindingsIndex.addUnsupportedTypealiasToSeen(usr);
   } else if (s is BooleanType) {
     // Ignore typedefs to Boolean.
-    logger.fine("Skipped Typedef '$typedefName': typedef to bool.");
-    bindingsIndex.addUnsupportedTypealiasToSeen(typedefUsr);
+    logger.fine("Skipped Typedef '$name': typedef to bool.");
+    bindingsIndex.addUnsupportedTypealiasToSeen(usr);
   } else {
     // Create typealias.
-    return Typealias(
-      usr: typedefUsr,
-      originalName: typedefName,
+    final type = Typealias(
+      usr: usr,
+      originalName: name,
       name: config.typedefs.rename(decl),
       type: s,
       dartDoc: getCursorDocComment(context, cursor),
     );
+    bindingsIndex.addTypealiasToSeen(usr, type);
+    return type;
   }
   return null;
 }