+++ pkgs/ffigen/CHANGELOG.md
@@ -10,6 +10,12 @@
   cannot collide with real USRs. It's very unlikely that any user facing USRs
   are affected.
 - Enum struct members now have setters.
+- __Breaking change__: Dart const values will be generated for global variables
+  marked const in C (e.g. static const int) instead of symbol lookups. This
+  supports integers, doubles, and string literals. Including the variable name
+  in the globals -> symbol-address configuration will still generate symbol 
+  lookups.
+
 
 ## 20.1.1